Разбил js файл на модули
